Update type on migration Add fucntions to get user liste

+++ migrations/2019-11-21-212329_create_users/up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users (
   id SERIAL PRIMARY KEY,
-  discordid INTEGER NOT NULL,
+  discordid BIGINT NOT NULL,
   role VARCHAR NOT NULL
 )
